+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/service/RoadAddressesService.scala
@@ -13,17 +13,13 @@ import org.slf4j.LoggerFactory
 
 case class ChangedRoadAddress(roadAddress : RoadAddress, link: RoadLink)
 
-class RoadAddressesService(val eventbus: DigiroadEventBus, roadLinkServiceImplementation: RoadLinkService) {
+class RoadAddressesService(val eventbus: DigiroadEventBus, roadLinkServiceImplementation: RoadLinkService ) {
 
-  private val roadNumberPublicId = "tie"          // Tienumero
-  private val roadPartNumberPublicId = "osa"      // Tieosanumero
-  private val startMeasurePublicId = "aet"        // Etaisyys
-  private val trackCodePublicId = "ajr"           // Ajorata
-  private val sideCodePublicId = "puoli"
+
 
   val roadAddressDAO = new RoadAddressDAO()
   val logger = LoggerFactory.getLogger(getClass)
-  val geometryTransform = new GeometryTransform
+//  val geometryTransform = new GeometryTransform
 
   def withDynTransaction[T](f: => T): T = OracleDatabase.withDynTransaction(f)
   def withDynSession[T](f: => T): T = OracleDatabase.withDynSession(f)
@@ -69,18 +65,5 @@ class RoadAddressesService(val eventbus: DigiroadEventBus, roadLinkServiceImplem
     }
   }
 
-  def getRoadAddressPropertiesByLinkId(persistedStop: PersistedMassTransitStop, roadLink: RoadLinkLike, oldProperties: Seq[Property]): Seq[Property] = {
-    val (address, roadSide) = geometryTransform.resolveAddressAndLocation(Point(persistedStop.lon, persistedStop.lat), persistedStop.bearing.get, persistedStop.mValue, persistedStop.linkId, persistedStop.validityDirection.get)
 
-    val newRoadAddressProperties =
-      Seq(
-        Property(0, roadNumberPublicId, PropertyTypes.ReadOnlyNumber, values = Seq(PropertyValue(address.road.toString, Some(address.road.toString)))),
-        Property(0, roadPartNumberPublicId, PropertyTypes.ReadOnlyNumber, values = Seq(PropertyValue(address.roadPart.toString, Some(address.roadPart.toString)))),
-        Property(0, startMeasurePublicId, PropertyTypes.ReadOnlyNumber, values = Seq(PropertyValue(address.addrM.toString, Some(address.addrM.toString)))),
-        Property(0, trackCodePublicId, PropertyTypes.ReadOnlyNumber, values = Seq(PropertyValue(address.track.value.toString, Some(address.track.value.toString)))),
-        Property(0, sideCodePublicId, PropertyTypes.ReadOnlyNumber, values = Seq(PropertyValue(roadSide.value.toString, Some(roadSide.value.toString))))
-      )
-
-    oldProperties.filterNot(op => newRoadAddressProperties.map(_.publicId).contains(op.publicId)) ++ newRoadAddressProperties
-  }
 }
